chore: refactor with better compliances on validations and code reuse
