+++ spid_cie_oidc/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.4.8"
+__version__ = "0.4.9"
